+++ lib/Controller/ApiController.php
@@ -213,7 +213,7 @@ class ApiController extends Controller {
 		$fileIds = \array_map('intval', \explode(',', $files));
 		$finalize = \filter_var($finalize, FILTER_VALIDATE_BOOLEAN);
 
-		$filesScanned = $this->scanner->scanFiles($this->userId, $this->userFolder, $fileIds);
+		$filesScanned = $this->scanner->scanFiles($this->userId, $fileIds);
 
 		$albumCoversUpdated = false;
 		if ($finalize) {